Tidy up ConversationEmail a bit
Just pass the email and avatar store in to the ctor, so they don't need
to be passed in and carried around all over the place. Ensure email
object is updated as more fields are loaded, so things like view source
still work. Ensure message progress pulsing stops on remote load error.