Add Email.Field.FIELDS to ImapDb.Folder.REQUIRED_FIELDS
This is especially needed since when moving a message from one folder to
another, the copy's flags are will un-flag it as deleted, so it actually
shows up again in conversations after the MoveEmailCommit op deletes the
original in the source folder.

This also has the extra benefit however of makig sure the engine knows
up-front what the uread state of all new mail is, and hence might
actaully get unread counts correct.