Fix FillWindowOperation not expanding vector when at end of it
This ensure that if a call to MinimalFolder.list_email_by_id() includes
a non-null initial_id, that its UID is always looked up. This fixes a
bug where if it is called with an initial ID, going newest to oldest,
and the id is the earliest email in the vector, no expansion will take
place.

This was occurring when scrolling the conversation list to the bottom,
and the ConversationMonitor had already reached the bottom of the
vector, so the fill op was trying to expand it. Since the last message
in the vector was given as the initial_id, but the fill op does not set
INCLUDE_ID, the UID for that email was never found. That caused the
vector expansion to use the end of the vector as the upper UID instead
of the this email at the start of the vector, and hence the lower UID
of the expansion was usually after the start of the vector, since
there's usally more messages in the vector than the fill op was
requesting. Thus no vector expansion actually took place, since the
lower and upper bounds were already included within the vector, and
hence no new messages were loaded by the fill op.

Fixes #219