Fix GOA accounts throwing an error when starting offline
Don't check service credentials are loaded when the account is being
opened, since GOA accounts will do a
org.gnome.OnlineAccounts.Account::EnsureCredentials() call, and this
will error out if the creds are invalid or not available.

Instead, check when we know that the services should be reachable, and
hope that means the GOA provider's auth service is reachable, too.