Fix some circular references on application startup
Since Applicaton.Controller constructs a MainWindow instance when it
is constructed, but the MainWindow and some other related classes
accesses the controller via the deprecated GearyApplication singleton
when they are cunstructed, the controller instance is still null and
a critical is reported instead.

Most of the code in question was related to classes needing to
know when the controller's account/folder/conversation monitor has
changed, so invert the call chain and have the controller let these
classes know.

The main culprit was the current conversation monitor, but since that
is very MainWindow-specific, just move it to the main window from the
controller, so the controller only needs to let the main window know
that the selected folder has changed.