Decouple Application.Contact from Geary.Contact
Update Application.Contact to have specific constructors for both
Folks-based and engine-based construction, remove the requirement for
the class to have a Geary.Contact field, and add a set of email
addresses as a property so that things like setting remote resource
loading prefs still works. This allows application contact to be
constructed from a Folks individual directly, without having to also
load an engine contact.

Update Application.ContactStore to cache Geary.Contact instances now
that they are always loaded from the database, rework loading code to
take advantage of the changes above.