Fix a deadlock resulting in no UI being shown on first run
After switching to DBus activation, we need to ensure that a controller
is present when arbitrary GApplication actions are executed so a
main window also exists. This lead to a deadlock on first run since
when the application was activated, it would be in the middle of
constructing the controller when it showed the Accounts Editor, which
would block since it would also try to construct a controller.

This fixes the deadlock by moving the first-run check to the
application, outside of the lock.