Decouple account processor closing and its op cancellables
By using a single cancellable for both managing the running state of
the AccountProcessor and its operations, if an op cancelled its own
cancellable, it would also stop the processor (and if it didn't, also
pre-cancel andy future ops).

Fix this by decopling processor running state from the operation
cancellables, and use a new cancellable for each operation.