Clean up composer handling on application quit
Rather than a convoluted systems of signals and callbacks to be able to
prompt if unsaved composers should be closed on application quit, just
do so directly from GearyApplication::quit, and close any still-open
composers on controller shutdown.

This requires some more complicated state handling in the composer, but
allows greatly simplifiying the processes to both quit the application
and close th controller.