ci: Don't build BlueZ backend
We figured out in libfolks that we were accidentally not properly
requiring the `python3-dbusmock` module that is necessary for the BlueZ
tests. We might want to be smarter about this in the future, but let's
just lose the offending backend for now.

Fixes https://gitlab.gnome.org/GNOME/geary/issues/695