Geary.Logging: Update how log blacklisting and suppression works
Blacklisting now prevents records from being created, where as
suppression now simply prevents them from being printed.

This means records are created for all log messages, regardless of
whether they are displayed.

Update API docs for a bunch of methods to make this clear, and to make
the requirements for getting log records in the first place is also
clear.