Plugin.EmailStore: Add methods for getting email ids from variants
Implement these in Application.EmailStoreFactory's implementation of
the interface.

This switches the variant format used to `(account_id,email_id)` to
match that used for folders, and so a folder doesn't need to be
specified when showing an email.

This breaks opening email from desktop notifications, but that will be
fixed shortly.