preferences: Set default_width to omit back button
The sizing of the preferences window depends on the button labels and
tends to be too narrow which makes the "back" button of
`HdyPreferencesWindow` to be shown. It's a bit weird to see this on the
desktop when there's plenty of space available.

This is shown by `HdyHeaderBar` when it considers a window to be
"mobile" which happens when its width is less than MOBILE_WINDOW_WIDTH
(400) or it's height less than MOBILE_WINDOW_HEIGHT (800) (see
`libhandy/src/hdy-header-bar.c`).